requirements.txt
pytorch-gpu>=1.10
numpy
scipy
scikit-learn
